gpreport/requirements.txt
beautifulsoup4==4.9.3
numpy==1.19.5
pandas==1.2.1
tqdm==4.56.0
pip==19.2.3
bump2version==0.5.11
wheel==0.33.6
watchdog==0.9.0
flake8==3.7.8
tox==3.14.0
coverage==4.5.4
Sphinx==1.8.5
twine==1.14.0
sqlalchemy==1.3.23
pathlib==1.0.1
pydantic==1.7.3
python-dotenv==0.15.0
psycopg2==2.8.6
schedule==1.0.0

